Fixed stuff I hope
I ran into a problem with the charts. I tried to fix it and actually have the fills be words, but then it gave me fatal error after fatal error so uh. I guess they are just gonna stay like that and I edited the ReadMe to be correct... I believe? I've mainly spent the last couple days studying for the exam so I think this one is gonna be rather subpar.